@@ -1,15 +1,14 @@
-import React from "react"
-import { Global, css } from "@emotion/react"
-import Layout from "./components/layout"
-import SideBar from "./components/sidebar/side-bar"
-import TabletBar from "./components/sidebar/tablet-bar"
-import { Route, Switch } from "react-router-dom"
-import AllMessages from "./page/all-messages"
-import MessageCRUD from "./service/message"
+import { Global, css } from "@emotion/react";
+import Layout from "./components/layout";
+import SideBar from "./components/sidebar/side-bar";
+import TabletBar from "./components/sidebar/tablet-bar";
+import { Routes, Route } from "react-router-dom"; // BrowserRouter 제거
+import AllMessages from "./page/all-messages";
+import MessageCRUD from "./service/message";
 
 export type AppProps = {
-  messageService: MessageCRUD
-}
+  messageService: MessageCRUD;
+};
 
 const App = ({ messageService }: AppProps) => {
   return (
@@ -19,18 +18,22 @@ const App = ({ messageService }: AppProps) => {
           <TabletBar />
           <SideBar />
         </Layout.Side>
-        <Switch>
-          <Layout.Main>
-            <Route exact path="/">
-              <AllMessages messageService={messageService} />
-            </Route>
-          </Layout.Main>
-        </Switch>
+        <Routes>
+          <Route
+            index
+            element={
+              <Layout.Main>
+                <AllMessages messageService={messageService} />
+              </Layout.Main>
+            }
+          />
+          {/* 필요하면 여기 다른 Route도 추가 */}
+        </Routes>
       </Layout>
       <Global styles={globalStyle} />
     </>
-  )
-}
+  );
+};
 
 const globalStyle = css`
   html,
@@ -48,6 +51,6 @@ const globalStyle = css`
       box-sizing: inherit;
     }
   }
-`
+`;
 
-export default App
+export default App;
