@@ -6,7 +6,7 @@ import "express-async-error";
 import messageRoute from "./routes/message";
 
 const app = express();
-const PORT = process.env.PORT || 3000;
+const PORT = process.env.PORT || 8080;
 
 app.use(express.json());
 app.use(helmet());
